% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_wrappers.R
\name{get_abundance_info}
\alias{get_abundance_info}
\title{get_abundance_info}
\usage{
get_abundance_info(
  sce,
  sample_id,
  group_id,
  celltype_id,
  min_cells = 10,
  covariates = NA
)
}
\arguments{
\item{sce}{SingleCellExperiment object of the scRNAseq data of interest.}

\item{sample_id}{Name of the meta data column that indicates from which sample/patient a cell comes from (in sce)}

\item{group_id}{Name of the meta data column that indicates from which group/condition a cell comes from (in sce)}

\item{celltype_id}{Name of the column in the meta data of sce that indicates the cell type of a cell.}

\item{min_cells}{Indicates the minimal number of cells that a sample should have to be considered in the DE analysis. Default: 10. See `muscat::pbDS`.}

\item{covariates}{NA if no covariates should be corrected for. If there should be corrected for covariates, this argument should be the name(s) of the columns in the meta data that indicate the covariate(s).}
}
\value{
List containing cell type abundance plots and the underlying data frame that was used to make these plots.
}
\description{
\code{get_abundance_info} Visualize cell type abundances. Calculate relative abundances of cell types as well.
}
\examples{
\dontrun{
library(dplyr)
sample_id = "tumor"
group_id = "pEMT"
celltype_id = "celltype"
abundance_info = get_abundance_info(sce = sce, sample_id = sample_id, group_id = group_id, celltype_id =  celltype_id, min_cells = 10)
}

}
