% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_wrappers.R
\name{get_expression_info}
\alias{get_expression_info}
\title{get_expression_info}
\usage{
get_expression_info(sce, sample_id, group_id, celltype_id, covariates = NA)
}
\arguments{
\item{sce}{SingleCellExperiment object of the scRNAseq data of interest.}

\item{sample_id}{Name of the meta data column that indicates from which sample/patient a cell comes from (in sce)}

\item{group_id}{Name of the meta data column that indicates from which group/condition a cell comes from (in sce)}

\item{celltype_id}{Name of the column in the meta data of sce that indicates the cell type of a cell.}

\item{covariates}{NA if no covariates should be corrected for. If there should be corrected for covariates, this argument should be the name(s) of the columns in the meta data that indicate the covariate(s).}
}
\value{
List containing data frames with average and fraction of expression per sample and per group.
}
\description{
\code{get_expression_info} Calculate the average and fraction of expression of each gene per sample and per group. Under the hood, the following functions are used: `get_avg_frac_exprs_abund`.
}
\examples{
\dontrun{
library(dplyr)
sample_id = "tumor"
group_id = "pEMT"
celltype_id = "celltype"
celltype_info = get_expression_info(sce = sce, sample_id = sample_id, group_id = group_id, celltype_id =  celltype_id)
}

}
