% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{muscat_analysis}
\alias{muscat_analysis}
\title{muscat_analysis}
\usage{
muscat_analysis(
sce, celltype_id, sample_id, group_id, covariates, contrasts_oi, contrast_tbl, assay_oi_pb ="counts", fun_oi_pb = "sum", de_method_oi = "edgeR", min_cells = 10, verbose = FALSE
)
}
\arguments{
\item{sce}{SingleCellExperiment object of the scRNAseq data of interest.}

\item{celltype_id}{Name of the column in the meta data of sce that indicates the cell type of a cell.}

\item{sample_id}{Name of the meta data column that indicates from which sample/patient a cell comes from (in sce)}

\item{group_id}{Name of the meta data column that indicates from which group/condition a cell comes from (in sce)}

\item{covariates}{NA if no covariates should be corrected for. If there should be corrected for covariates, this argument should be the name(s) of the columns in the meta data that indicate the covariate(s).}

\item{contrasts_oi}{String indicating the contrasts of interest (= which groups/conditions will be compared) for the differential expression and MultiNicheNet analysis.
We will demonstrate here a few examples to indicate how to write this. Check the limma package manuals for more information about defining design matrices and contrasts for differential expression analysis.\cr
If wanting to compare group A vs B: `contrasts_oi = c("'A-B'")` \cr
If wanting to compare group A vs B & B vs A: `contrasts_oi = c("'A-B','B-A'")` \cr
If wanting to compare group A vs B & A vs C & A vs D: `contrasts_oi = c("'A-B','A-C', 'A-D'")` \cr
If wanting to compare group A vs B and C: `contrasts_oi = c("'A-(B+C)/2'")` \cr
If wanting to compare group A vs B, C and D: `contrasts_oi = c("'A-(B+C+D)/3'")` \cr
If wanting to compare group A vs B, C and D & B vs A,C,D: `contrasts_oi = c("'A-(B+C+D)/3', 'B-(A+C+D)/3'")` \cr
Note that the groups A, B, ... should be present in the meta data column 'group_id'.}

\item{contrast_tbl}{Data frame providing names for each of the contrasts in contrasts_oi in the 'contrast' column, and the corresponding group of interest in the 'group' column. Entries in the 'group' column should thus be present in the group_id column in the metadata.
Example for `contrasts_oi = c("'A-(B+C+D)/3', 'B-(A+C+D)/3'")`:
`contrast_tbl = tibble(contrast = c("A-(B+C+D)/3","B-(A+C+D)/3"), group = c("A","B"))`}

\item{assay_oi_pb}{Indicates which information of the assay of interest should be used (counts, scaled data,...). Default: "counts". See `muscat::aggregateData`.}

\item{fun_oi_pb}{Indicates way of doing the pseudobulking. Default: "sum". See `muscat::aggregateData`.}

\item{de_method_oi}{Indicates the DE method that will be used after pseudobulking. Default: "edgeR". See `muscat::pbDS`.}

\item{min_cells}{Indicates the minimal number of cells that a sample should have to be considered in the DE analysis. Default: 10. See `muscat::pbDS`.}

\item{verbose}{Indicate which different steps of the pipeline are running or not. Default: FALSE.}
}
\value{
List containing information and output of the Muscat analysis.\cr
celltype_info: contains average expression value and fraction of each cell type - sample combination,\cr
celltype_de: contains output of the differential expression analysis,\cr
grouping_tbl: data frame showing the group per sample \cr
}
\description{
\code{muscat_analysis}  Perform a multi-sample multi-condition DE analysis with the pseudobulk approach implemented in muscat.
}
\examples{
\dontrun{
library(dplyr)
sample_id = "tumor"
group_id = "pEMT"
celltype_id = "celltype"
covariates = NA
contrasts_oi = c("'High-Low','Low-High'")
contrast_tbl = tibble(contrast = c("High-Low","Low-High"), group = c("High","Low"))
output = muscat_analysis(
     sce = sce,
     celltype_id = celltype_id,
     sample_id = sample_id,
     group_id = group_id,
     covariates = covariates,
     contrasts_oi = contrasts_oi,
     contrast_tbl = contrast_tbl)
}

}
